Print status output as a dense table
Implements logger handling the status structure to make it clean and
dense and to handle when the window isn't wide enough for the output
provided.

The missing piece is to provide more verbose logging at checkpoints to
provide extra details and a history of the behavior in the log output.

Refs TS-4180

Change-Id: I47079bd59b8eb6f13c5ecb9daa1ed9a204f5c9d8
Reviewed-on: https://gerrit.instructure.com/152278
Tested-by: Jenkins
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
Reviewed-by: Mark Severson <markse@instructure.com>
Product-Review: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Ryan Taylor <rtaylor@instructure.com>